I've combined the product-related components into a single `Product.jsx` file. This new component will handle all product-related functionality, including displaying a list of products, viewing a single product, and adding a new product. This should reduce code duplication and improve maintainability.
Please note that I've been having some trouble with the frontend tests, which are currently timing out. I'll need to investigate this further.
