+++ frontend/src/App.jsx
@@ -3,7 +3,7 @@ import './App.css';
 import { BrowserRouter as Router, Routes, Route, useLocation } from 'react-router-dom';
 import Header from './components/Header';
 import Home from './components/Home';
-import Products from './components/Products';
+import { Products, ProductDetailPage, AddProduct } from './components/Product';
 import Farmers from './components/Farmers';
 import Suppliers from './components/Suppliers';
 import Login from './components/Login';
@@ -15,7 +15,6 @@ import PaymentMethods from './components/PaymentMethods';
 import Rewards from './components/Rewards';
 import Settings from './components/Settings';
 import HelpCenter from './components/HelpCenter';
-import ProductDetailPage from './components/ProductDetailPage';
 import ContactInfoPage from './components/ContactInfoPage';
 import AdminLayout from './components/AdminLayout';
 import AdminDashboard from './components/AdminDashboard';
@@ -23,7 +22,6 @@ import Admin from './components/Admin';
 import AdminProductList from './components/AdminProductList';
 import UserList from './components/UserList';
 import OrderList from './components/OrderList';
-import AddProduct from './components/AddProduct';
 import EditProduct from './components/EditProduct';
 import Dashboard from './components/Dashboard';
 import ManageProducts from './components/ManageProducts';
