--- frontend/src/components/AddProduct.jsx
@@ -1,169 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import axios from '../api/axios';
-import { useNavigate } from 'react-router-dom';
-import '../styles/Auth.css';
-
-const AddProduct = () => {
-  const [formData, setFormData] = useState({
-    name: '',
-    description: '',
-    price: '',
-    quantity_available: '',
-    category: '',
-    images: [],
-  });
-  const [categories, setCategories] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState('');
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    const fetchCategories = async () => {
-      try {
-        const response = await axios.get('products/categories/');
-        setCategories(response.data.results);
-      } catch (err) {
-        console.error('Error fetching categories:', err);
-      }
-    };
-    fetchCategories();
-  }, []);
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setFormData({ ...formData, [name]: value });
-  };
-
-  const handleImageChange = (e) => {
-    setFormData({ ...formData, images: e.target.files });
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setLoading(true);
-    setError('');
-
-    const productData = new FormData();
-    productData.append('name', formData.name);
-    productData.append('description', formData.description);
-    productData.append('price', formData.price);
-    productData.append('quantity_available', formData.quantity_available);
-    productData.append('category', formData.category);
-
-    for (let i = 0; i < formData.images.length; i++) {
-      productData.append('uploaded_images', formData.images[i]);
-    }
-
-    try {
-      await axios.post('products/', productData, {
-        headers: {
-          'Content-Type': 'multipart/form-data',
-          Authorization: `Bearer ${localStorage.getItem('access_token')}`,
-        },
-      });
-      navigate('/products');
-    } catch (err) {
-      console.error("Error adding product:", err);
-      if (err.response) {
-        console.error("Error data:", err.response.data);
-        console.error("Error status:", err.response.status);
-        console.error("Error headers:", err.response.headers);
-      } else if (err.request) {
-        console.error("Error request:", err.request);
-      } else {
-        console.error('Error', err.message);
-      }
-      setError('There was an error adding the product. Please check the console for more details.');
-      console.error("Error adding product:", err.response ? err.response.data : err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="auth-page">
-      <div className="auth-container">
-        <h2>Add a New Product</h2>
-        {error && <div className="error-message">{error}</div>}
-        <form onSubmit={handleSubmit}>
-          <div className="form-group">
-            <label htmlFor="name">Product Name</label>
-            <input
-              type="text"
-              id="name"
-              name="name"
-              value={formData.name}
-              onChange={handleChange}
-              required
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="description">Description</label>
-            <textarea
-              id="description"
-              name="description"
-              value={formData.description}
-              onChange={handleChange}
-              required
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="price">Price</label>
-            <input
-              type="number"
-              id="price"
-              name="price"
-              value={formData.price}
-              onChange={handleChange}
-              required
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="quantity_available">Quantity Available</label>
-            <input
-              type="number"
-              id="quantity_available"
-              name="quantity_available"
-              value={formData.quantity_available}
-              onChange={handleChange}
-              required
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="category">Category</label>
-            <select
-              id="category"
-              name="category"
-              value={formData.category}
-              onChange={handleChange}
-              required
-            >
-              <option value="">Select a category</option>
-              {categories.map((category) => (
-                <option key={category.id} value={category.id}>
-                  {category.name}
-                </option>
-              ))}
-            </select>
-          </div>
-          <div className="form-group">
-            <label htmlFor="images">Product Images</label>
-            <input
-              type="file"
-              id="images"
-              name="images"
-              onChange={handleImageChange}
-              multiple
-              required
-            />
-          </div>
-          <button type="submit" className="auth-btn" disabled={loading}>
-            {loading ? 'Adding...' : 'Add Product'}
-          </button>
-        </form>
-      </div>
-    </div>
-  );
-};
-
-export default AddProduct;
